[ingest]: move NewClientForTesting to pkg/ingest/testutil.go
Separates the test helper from the production client code. No behaviour
change; call sites remain ingest.NewClientForTesting(...).
