+++ pkg/ingest/reader_client.go
@@ -145,15 +145,6 @@ func (c *Client) Close() {
 	c.Client.Close() // Close the underlying client
 }
 
-// NewClientForTesting wraps a *kgo.Client in an ingest.Client without starting
-// the partition monitor goroutine. For use in unit tests only.
-func NewClientForTesting(client *kgo.Client) *Client {
-	return &Client{
-		Client: client,
-		stopCh: make(chan struct{}),
-		logger: log.NewNopLogger(),
-	}
-}
 
 func NewReaderClientMetrics(component string, reg prometheus.Registerer) *kprom.Metrics {
 	return kprom.NewMetrics("tempo_ingest_storage_reader",
